+++ src/main/java/org/truffleruby/language/loader/FeatureLoader.java
@@ -12,6 +12,8 @@ package org.truffleruby.language.loader;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -107,7 +109,7 @@ public class FeatureLoader {
         try {
             final Map<String, List<RubyConstant>> constantsMap = registeredAutoloads.get(basename);
             if (constantsMap == null || constantsMap.isEmpty()) {
-                return null;
+                return Collections.emptyList();
             }
 
             // Deep-copy constantsMap so we can call findFeature() outside the lock
@@ -124,20 +126,11 @@ public class FeatureLoader {
             final String expandedAutoloadPath = findFeature(entry.getKey());
 
             if (expandedPath.equals(expandedAutoloadPath)) {
-                for (RubyConstant constant : entry.getValue()) {
-                    // Do not autoload recursively from the #require call in GetConstantNode
-                    if (!constant.getAutoloadConstant().isAutoloading()) {
-                        constants.add(constant);
-                    }
-                }
+                constants.addAll(Arrays.asList(entry.getValue()));
             }
         }
 
-        if (constants.isEmpty()) {
-            return null;
-        } else {
-            return constants;
-        }
+        return constants;
     }
 
     public void removeAutoload(RubyConstant constant) {