+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Compatibility:
 * Update `Range#cover?` to handle `Range` parameter.
 * Fix `String#{casecmp, casecmp?}` parameter conversion.
 * Fix `Regexp` issue which raised syntax error instead of `RegexpError` (#2066).
+* Handle `Object#autoload` when autoload itself (#1616, @ssnickolay)
 
 Performance: